# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/behavior/class-progress/{class_id}": {
            "get": {
                "description": "获取学生在某个班级下的所有视频学习情况",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "behavior"
                ],
                "summary": "获取班级内所有视频的观看进度",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "班级ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/behavior.GetClassVideoProgressResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/behavior/record": {
            "post": {
                "description": "记录暂停、回放、播放等行为及时长",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "behavior"
                ],
                "summary": "记录学生视频观看行为",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "行为上报参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/behavior.RecordActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/behavior/update-progress": {
            "post": {
                "description": "上报当前播放秒数，更新最大进度与完成状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "behavior"
                ],
                "summary": "更新学生视频观看进度",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "进度参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/behavior.UpdateProgressReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "behavior.GetClassVideoProgressResp": {
            "type": "object",
            "properties": {
                "progress_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/behavior.VideoProgress"
                    }
                }
            }
        },
        "behavior.RecordActionReq": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "event": {
                    "description": "pause|replay",
                    "type": "string"
                },
                "segment_id": {
                    "type": "integer"
                },
                "video_id": {
                    "type": "integer"
                }
            }
        },
        "behavior.UpdateProgressReq": {
            "type": "object",
            "properties": {
                "current_sec": {
                    "type": "integer"
                },
                "video_id": {
                    "type": "integer"
                }
            }
        },
        "behavior.VideoProgress": {
            "type": "object",
            "properties": {
                "progress_percent": {
                    "description": "完成进度百分比",
                    "type": "integer"
                },
                "status": {
                    "description": "finished | in_progress | todo",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "video_id": {
                    "type": "integer"
                },
                "watch_time": {
                    "description": "总观看时长",
                    "type": "integer"
                }
            }
        },
...
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}